# platform/docs/skill-integration/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Algoria API",
    "version": "1.0.0",
    "description": "Own service catalogue and HTTP APIs, Stellar testnet x402 payments, sync or async execution. Discovery metadata is served locally; it is not published to an external catalogue."
  },
  "servers": [
    {
      "url": "https://vqqbvydiehuwdzbgvmun.supabase.co/functions/v1/api"
    }
  ],
  "paths": {
...
    "/discovery/resources": {
      "get": {
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Exact-match filter within our own service catalogue."
          },
          {
            "name": "network",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Exact-match filter within our own service catalogue."
          },
          {
            "name": "scheme",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Exact-match filter within our own service catalogue."
          },
          {
            "name": "payTo",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Exact-match filter within our own service catalogue."
          },
          {
            "name": "extensions",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Exact-match filter within our own service catalogue."
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Own service catalogue",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "x402Version",
                    "resources",
                    "pagination"
                  ],
                  "properties": {
                    "x402Version": {
                      "const": 2
                    },
                    "resources": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "version",
                          "name",
                          "description",
                          "tags",
                          "type",
                          "x402Version",
                          "resource",
                          "method",
                          "input_schema",
                          "output_schema",
                          "execution",
                          "accepts",
                          "extensions",
                          "headers",
                          "responses",
                          "recovery"
                        ],
                        "properties": {
                          "id": {
                            "const": "image.generate"
                          },
                          "version": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "tags": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "type": {
                            "const": "http"
                          },
                          "x402Version": {
                            "const": 2
                          },
                          "resource": {
                            "type": "string",
                            "format": "uri"
                          },
                          "method": {
                            "const": "POST"
                          },
                          "input_schema": {
                            "type": "object"
                          },
                          "output_schema": {
                            "type": "object"
                          },
                          "execution": {
                            "type": "object",
                            "required": [
                              "supported",
                              "default",
                              "default_wait_ms",
                              "max_wait_ms"
                            ],
                            "properties": {
                              "supported": {
                                "type": "array",
                                "items": {
                                  "enum": [
                                    "sync",
                                    "async"
                                  ]
                                }
                              },
                              "default": {
                                "const": "sync"
                              },
                              "default_wait_ms": {
                                "type": "integer"
                              },
                              "max_wait_ms": {
                                "type": "integer"
                              }
                            }
                          },
                          "accepts": {
                            "type": "array",
                            "minItems": 1,
                            "items": {
                              "type": "object",
                              "required": [
                                "scheme",
                                "network",
                                "asset",
                                "amount",
                                "payTo",
                                "maxTimeoutSeconds",
                                "extra"
                              ],
                              "properties": {
                                "scheme": {
                                  "const": "exact"
                                },
                                "network": {
                                  "const": "stellar:testnet"
                                },
                                "asset": {
                                  "type": "string"
                                },
                                "amount": {
                                  "type": "string",
                                  "pattern": "^[1-9]\\d*$"
                                },
                                "payTo": {
                                  "type": "string"
                                },
                                "maxTimeoutSeconds": {
                                  "type": "integer",
                                  "minimum": 1
                                },
                                "extra": {
                                  "type": "object",
                                  "required": [
                                    "areFeesSponsored"
                                  ],
                                  "properties": {
                                    "areFeesSponsored": {
                                      "const": true
                                    }
                                  }
                                }
                              }
                            }
                          },
                          "extensions": {
                            "type": "object",
                            "required": [
                              "bazaar"
                            ],
                            "properties": {
                              "bazaar": {
                                "type": "object",
                                "required": [
                                  "info",
                                  "schema"
                                ],
                                "properties": {
                                  "info": {
                                    "type": "object"
                                  },
                                  "schema": {
                                    "type": "object"
                                  }
                                }
                              }
                            }
                          },
                          "headers": {
                            "type": "object",
                            "additionalProperties": {
                              "type": "string"
                            }
                          },
                          "responses": {
                            "type": "object",
                            "additionalProperties": {
                              "type": "string"
                            }
                          },
                          "recovery": {
                            "type": "object",
                            "required": [
                              "method",
                              "url_template",
                              "authorization",
                              "polling_interval_ms"
                            ],
                            "properties": {
                              "method": {
                                "const": "GET"
                              },
                              "url_template": {
                                "type": "string"
                              },
                              "authorization": {
                                "type": "string"
                              },
                              "polling_interval_ms": {
                                "type": "integer"
                              }
                            }
                          }
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "required": [
                        "limit",
                        "offset",
                        "total",
                        "cursor"
                      ],
                      "properties": {
                        "limit": {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 100
                        },
                        "offset": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "total": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "cursor": {
                          "type": "null"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Temporarily unavailable; retry the same identity or check status.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/jobs/{job_id}": {
      "get": {
        "security": [
          {
            "recoveryToken": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current job state, receipt and result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "404": {
            "description": "Missing job or invalid recovery token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Temporarily unavailable; retry the same identity or check status.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "job_id": {
            "type": "string",
            "format": "uuid"
          },
          "payment": {
            "type": "object",
            "required": [
              "success",
              "network",
              "transaction"
            ],
            "properties": {
              "success": {
                "type": "boolean"
              },
              "network": {
                "const": "stellar:testnet"
              },
              "transaction": {
                "type": "string"
              },
              "payer": {
                "type": "string"
              },
              "amount": {
                "type": "string",
                "pattern": "^\\d+$"
              },
              "errorReason": {
                "type": "string"
              },
              "errorMessage": {
                "type": "string"
              },
              "extensions": {
                "type": "object"
              },
              "extra": {
                "type": "object"
              }
            }
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "recoveryToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "The job recovery token created locally before the first request."
      }
    }
  }
}